% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplik.R
\name{emplik}
\alias{emplik}
\title{empirical likelihood method based on momentfit package}
\usage{
emplik(target, instr, data, theta_0, ...)
}
\arguments{
\item{target}{formula for target variable}

\item{instr}{a formula for covariate of response}

\item{data}{input data}

\item{theta_0}{initial moment condition for coeffs}
}
\description{
Importing
}
\author{
Maciej Ostapiuk and Maciej Beręsewicz based on Kim, Shao (2021)
}
