% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencal.R
\name{gencal}
\alias{gencal}
\title{Generalized calibration function based on sampling::gencalib}
\usage{
gencal(Xs, Zs, d, totals, method = "raking", eps, maxit, tol)
}
\arguments{
\item{Xs}{A vector of population totals}

\item{Zs}{A matrix of calibration variables}

\item{d}{A matrix of instrumental variables}

\item{totals}{A vector of initial weights, derived from sampling design}

\item{method}{TBA}

\item{eps}{TBA}

\item{maxit}{TBA}

\item{tol}{TBA}
}
\description{
Importing
}
\author{
Maciej Ostapiuk and Maciej Beręsewicz based on Kott and Chang (2010) and Kott and Liao (2017)
}
