% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnar.R
\name{mnar}
\alias{mnar}
\title{The main function for the not missing at random non-response}
\usage{
mnar(
  response,
  outcome,
  calibration,
  target,
  data,
  svydesign,
  dweights,
  N,
  pop_totals,
  method = c("gencalib", "emplik", "gmm"),
  tol = 1e-08,
  maxit = 50,
  eps = .Machine$double.eps,
  control = NULL,
  ...
)
}
\arguments{
\item{response}{response variables (instruments)}

\item{outcome}{outcome variables (for modelling E(Y|X))}

\item{calibration}{calibration variables (for which totals should be reproduced)}

\item{target}{target variable}

\item{data}{data.frame}

\item{svydesign}{surveydesign object}

\item{dweights}{design weights}

\item{N}{population size}

\item{pop_totals}{population totals (for calibration)}

\item{method}{methods \code{c("gencalib", "emplik", "gmm")}}

\item{tol}{tolerance \code{1e-8}}

\item{maxit}{maxit \code{50}}

\item{eps}{eps for inverse \code{.Machine$double.eps}}

\item{control}{control for methods}

\item{...}{TBA}
}
\value{
Returns a list with containing:\cr
\itemize{
\item{\code{g} -- g-weight that sums up to sample size,}
\item{\code{Xs} -- matrix used for calibration (i.e. Intercept, X and X_q transformed for calibration of quantiles),}
\item{\code{totals} -- a vector of totals (i.e. \code{N}, \code{pop_totals} and \code{pop_quantiles}),}
\item{\code{method} -- selected method,}
\item{\code{backend} -- selected backend.}
}
}
\description{
Here is the part where the description will go
}
\details{
Imports for the function
}
\examples{

set.seed(123)
n <- 1000

}
\references{
Deville, J. C., and Särndal, C. E. (1992). Calibration estimators in survey sampling.
Journal of the American statistical Association, 87(418), 376-382.

add others
}
\seealso{
\code{\link[sampling:calib]{sampling::calib()}} -- for standard calibration.

\code{\link[laeken:calibWeights]{laeken::calibWeights()}} -- for standard calibration.

\code{\link[survey:calibrate]{survey::calibrate()}} -- for standard and more advanced calibration.

\code{\link[ebal:ebalance]{ebal::ebalance()}} -- for standard entropy balancing.
}
\author{
Maciej Ostapiuk, Maciej Beręsewicz
}
