% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencal.R
\name{gencal}
\alias{gencal}
\title{An internal function for calibration of weights to adjust for nonignorable nonresponse}
\usage{
gencal(totals, calib_var, instr_var = NULL, target_var, initial_weights)
}
\arguments{
\item{totals}{A vector of population totals}

\item{calib_var}{A matrix of calibration variables}

\item{instr_var}{A matrix of instrumental variables}

\item{target_var}{A vector of target variables}

\item{initial_weights}{A vector of initial weights, derived from sampling design}
}
\description{
An internal function for calibration of weights to adjust for nonignorable nonresponse
}
\author{
Maciej Ostapiuk based on Kott and Chang (2010) and Kott and Liao (2017)
}
